backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pyyaml==6.0.2
pillow==11.0.0
pypdfium2==4.30.0
python-dotenv==1.0.1
httpx==0.27.2
vllm==0.6.3
